Context.getDir not persist sometimes in provider for user 0.
Symptom:
As issue link described:
https://code.google.com/p/android/issues/detail?id=184067

When start process for provider, if process not exists before,
the dataDir in ApplicationInfo will be set to /data/user/0/

But start process for service, the dataDir will be /data/data/
except userId is not zero.

Caused calling Context.getDir the prefix of data path is not same
in sometimes.

RootCause:
In getContentProviderImpl() of ActivityManagerService.java,
it use getAppInfoForUser() directly whether user ID is 0 or not.

In ActiveServices.java has the same condition,
will use getAppInfoForUser() if user ID is larger than 0.

Solution:
Call getAppInfoForUser when user ID larger than 0 as service case.

Change-Id: I2938fec6d5a4a472cc60046a1544152dbc9960c7
Signed-off-by: mydongistiny <jaysonedson@gmail.com>