Escalate HeadsUp when dreaming start
After locked with dreaming, pinned headsup can be remained
in Keyguard State. This is becuase device is locked without
screen off, and HeadsUp is not esclated.
As a result Pinned headsup is shown behind boucer when unlock.
In this case touchable bound is limited to headsup bound.
So, bouncer is not touchable at all.

Bug:30820511
Change-Id: I403b4a02da1dd3e70f02702da88fbd3a7bc9cbca