Fix keyguard flash issue
When we start camera from keyguard, if keyguard is secure,
this will lead to an abnormal flash when pressing back button.

https://code.google.com/p/android/issues/detail?id=228248

Solution:
When the new app window is visible and we know whether to show the
keyguard,we need to reset the keyguard status at first.

Change-Id: I10aa4d7e39bd659592f5726f5adbd17bfdf06dfc
Signed-off-by: longhai <longhai@xiaomi.com>
(cherry picked from commit 5660802d54814aa4a0854d79113b68785d5a2bb5)