QS Tiles: Add High Brightness Mode Tile
Add a tile to toggle kernel based High Brightness Mode.
This is a conditional tile and will only show if the proper support is found in FWB and device tree.

Proof of Concept tile. This should actually be incorporated as a conditional third state in the existing brightness tile to make it more apparent to users that the manual and auto brightness modes are overriden by this. Essentialy to indicate that it is a third mode.

Change-Id: I886d72b0e40bebeb3ed70bb6956ddd104842b7ed

# Conflicts:
#	packages/SystemUI/res/values/du_strings.xml